Refactor to classes, preparation for #1
